refactor: convert all IDs to Integer from Long
Update all entity models, repositories, services, controllers, DTOs and utility classes to use Integer instead of Long for all ID types
Fix database connection URL to use gppremium instead of gppremium3
Bump project version to 1.10.1
Clean up pom.xml by removing unused dependencies and adjust build plugin configurations
Update README.md to include WildFly build command and remove outdated hardcoded line
Fix missing trailing newlines in several source files
Update all test files to match the new ID type conventions

+++ src/main/java/br/compneusgppremium/api/config/security/AutenticacaoViaTokenFilter.java
@@ -38,7 +38,7 @@ public class AutenticacaoViaTokenFilter extends OncePerRequestFilter {
 
 	private void autenticarCliente(String token) {
 		try {
-			Long idUsuario = tokenService.getIdUsuario(token);
+			Integer idUsuario = tokenService.getIdUsuario(token);
 
 			repository.findById(idUsuario).ifPresent(usuarioModel -> {
 				UsernamePasswordAuthenticationToken authentication =
