+++ src/main/java/br/compneusgppremium/api/controller/dto/UsuarioResponsavelDTO.java
@@ -5,7 +5,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 @Schema(description = "DTO com informações do usuário responsável")
 public class UsuarioResponsavelDTO {
     @Schema(description = "ID do usuário", example = "1")
-    private Long id;
+    private Integer id;
     
     @Schema(description = "Nome do usuário", example = "Maria Santos")
     private String nome;
@@ -13,19 +13,19 @@ public class UsuarioResponsavelDTO {
     @Schema(description = "Login do usuário", example = "maria.santos")
     private String login;
 
-    public UsuarioResponsavelDTO(Long id, String nome) {
+    public UsuarioResponsavelDTO(Integer id, String nome) {
         this.id = id;
         this.nome = nome;
     }
 
-    public UsuarioResponsavelDTO(Long id, String nome, String login) {
+    public UsuarioResponsavelDTO(Integer id, String nome, String login) {
         this.id = id;
         this.nome = nome;
         this.login = login;
     }
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
 
     public String getNome() { return nome; }
     public void setNome(String nome) { this.nome = nome; }
